backend/config/requirements-prod.txt
alembic==1.18.4
blinker==1.9.0
click==8.3.1
Flask==3.1.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.2.2
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
SQLAlchemy==2.0.48
typing_extensions==4.15.0
Werkzeug==3.1.6

# Production dependencies
gunicorn==21.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9

# Africa's Talking SDK (add when ready)
# africastalking==1.2.7
